Consolidate git-backed build helpers in git utils
